Support variadic templates from STL
This commit adds the support for variadic templates
as long as they are coming from the standard library.
It was achieved by supporting the Pack type of
CXTemplateArgumentKind.
This commit also fixed a bug of getting the full
name of a CppElement so it does not throw
a lot of unexpected exceptions.