Default templated arguments are not working
If using the visitor pattern with libclang to check the list of function
arguments, it is impossible to differentiate default arguments from template
arguments. The reason is the libclang within its VisitTemplateSpecializationTypeLoc
function first sends back a TemplateRef with the actual template being used, right
after that it sends back the (some of the) template arguments. However the
template arguments are sent back as the same literal expressions, like a normal
=5 expression, so it is impossible to differentiate if the received expression
is actually a default argument or a template parameter.
To overcome this issue default argument handling for parameters are done
differently from this time on, as CXCursor has a built-in feature to fetch the
default argument for the cursor.